State machine controllers.
Switching focus between different parts of the application
increases the complexity of validating what actions are valid
or not. Having controllers return a controller on each action
allows to make invalid states not representable which simplifies
things a lot and is a design that scales well as the app grows.
This is a state machine design where each action returns the next
state and the set of applicable actions for the next state.